refactored useNft hooks
