+++ src/components/child/ParentInfo.vue
@@ -78,11 +78,11 @@
 <script lang="ts">
 import { getShortenAddress } from '@astar-network/astar-sdk-core';
 import ActionButtons from 'src/components/child/ActionButtons.vue';
-import { useBreakpoints, useNft, useNft2, useToken } from 'src/hooks';
+import { useBreakpoints, useToken } from 'src/hooks';
 import { ExtendedAsset, IBasePart, Id } from 'src/modules/nft';
 import { networkParam, Path } from 'src/router/routes';
 import { computed, defineComponent, ref, watch } from 'vue';
-import { useRoute, useRouter } from 'vue-router';
+import { useRouter } from 'vue-router';
 
 enum InventoryTab {
   inventory = 'Inventory',
@@ -129,15 +129,15 @@ export default defineComponent({
   setup(props) {
     const { width, screenSize } = useBreakpoints();
     const router = useRouter();
-    const route = useRoute();
 
-    const { unequip, equip, isLoading } = useNft(Number(props.parentTokenId));
-    const { token, fetchToken } = useToken(props.parentContractAddress, props.parentTokenId);
+    const { token, fetchToken, equip, unequip, isLoading } = useToken(
+      props.parentContractAddress,
+      props.parentTokenId
+    );
     const parts = computed(() => token.value?.assets[0].parts ?? []);
     const itemPreview = ref<[Id, (ExtendedAsset | null)[]]>();
     const isLoadingItem = ref<boolean>(false);
     const selectedTab = ref<InventoryTab>(InventoryTab.inventory);
-    const { getChildrenToEquipPreview } = useNft2();
 
     const loadParentToken = async (): Promise<void> => {
       await fetchToken(true);
@@ -176,21 +176,6 @@ export default defineComponent({
       return false;
     });
 
-    const setItemPreview = async (): Promise<void> => {
-      if (!isLoading.value && slotVacant.value) {
-        isLoadingItem.value = true;
-        const preview = await getChildrenToEquipPreview(
-          props.parentContractAddress,
-          Number(props.parentTokenId)
-        );
-        if (!preview)
-          throw Error(`there are no equitable child Nfts with parentId:${props.parentTokenId}`);
-        itemPreview.value = Array.from(preview).find(
-          (it) => it[0].u64 === Number(props.childTokenId)
-        );
-        isLoadingItem.value = false;
-      }
-    };
     const handleUnequip = async (): Promise<void> => {
       const part = parts.value.find(
         (it) =>
@@ -204,12 +189,11 @@ export default defineComponent({
     };
 
     const handleEquip = async (): Promise<void> => {
-      if (!itemPreview.value || !slotVacant.value) return;
+      if (!slotVacant.value) return;
       const slotId = Number(slotVacant.value.id);
       await equip(
         slotId,
-        itemPreview.value[0],
-        itemPreview.value[1],
+        props.childTokenId,
         props.parentContractAddress,
         props.childContractAddress
       );
@@ -222,8 +206,6 @@ export default defineComponent({
       const url = `${base}?parentId=${props.parentTokenId}&contractAddress=${props.parentContractAddress}`;
       router.push(url);
     };
-
-    watch([parts, isLoading], setItemPreview, { immediate: true });
 
     return {
       selectedTab,
